+++ config/me_youtube.php
@@ -4,12 +4,14 @@ return ['MeYoutube' => [
 	'video' => [
 		//"Show" options
 		'show' => [
-			//Displays the post author
+			//Displays the video author
 			'author' => TRUE,
-			//Displays the post category
+			//Displays the video category
 			'category' => TRUE,
-			//Displays the post created datetime
+			//Displays the video created datetime
 			'created' => TRUE,
+			//Plays a spot automatically before each video
+			'spot' => TRUE,
 			//Displays the Shareaholic social buttons
 			//Remember you have to set app and site IDs. See `shareaholic.app_id` and `shareaholic.site_id` in the MeCms configuration
 			'shareaholic' => FALSE